confocal: add BaseScan and ConfocalImage classes
