+++ lumicks/pylake/point_scan.py
@@ -1,25 +1,24 @@
 import json
 
-from .detail.mixin import PhotonCounts
-from .detail.mixin import ExcitationLaserPower
+from .confocal import BaseScan
 
 
-class PointScan(PhotonCounts, ExcitationLaserPower):
-    """A point scan exported from Bluelake
+class PointScan(BaseScan):
+    """A confocal point scan exported from Bluelake
 
     Parameters
     ----------
-    h5py_dset : h5py.Dataset
-        The original HDF5 dataset containing the point scan
+    name : str
+        point scan name
     file : lumicks.pylake.File
-        The parent file. Used to look up channel data.
+        Parent file. Contains the channel data.
+    start : int
+        Start point in the relevant info wave.
+    stop : int
+        End point in the relevant info wave.
+    json : dict
+        Dictionary containing scan-specific metadata.
     """
-    def __init__(self, name, file, start, stop, json):
-        self.start = start
-        self.stop = stop
-        self.name = name
-        self.json = json
-        self.file = file
 
     def _get_photon_count(self, name):
         return getattr(self.file, f"{name}_photon_count".lower())[self.start:self.stop]
@@ -82,20 +81,3 @@ class PointScan(PhotonCounts, ExcitationLaserPower):
         """
         for color in ["red", "green", "blue"]:
             self._plot_color(color, **kwargs)
-
-    @classmethod
-    def from_dataset(cls, h5py_dset, file):
-        """Construct PointScan class from dataset.
-
-        Parameters
-        ----------
-        h5py_dset : h5py.Dataset
-            The original HDF5 dataset containing kymo information
-        file : lumicks.pylake.File
-            The parent file. Used to loop up channel data
-        """
-        start = h5py_dset.attrs["Start time (ns)"]
-        stop = h5py_dset.attrs["Stop time (ns)"]
-        name = h5py_dset.name.split("/")[-1]
-        json_data = json.loads(h5py_dset[()])["value0"]
-        return cls(name, file, start, stop, json_data)
